AudioRecord: record using using AsyncTask
