+++ app/src/main/java/com/example/keshavdulal/a14_simple_drawing/GraphFragment.java
@@ -76,7 +76,7 @@ public class GraphFragment extends Fragment{
                 X2=X1;
                 Y2=Y1-graph_height;
                 canvas.drawLine(X1,Y1,X2,Y2,graphLinesObj);
-                graph_height = ((float)AudioRecordClass.temp)/50;
+                graph_height = ((float) MainActivity.temp)/50;
             }
             postInvalidateDelayed(250);
