+++ app/src/main/java/com/example/keshavdulal/a14_simple_drawing/MainActivity.java
@@ -2,7 +2,12 @@ package com.example.keshavdulal.a14_simple_drawing;
 
 
 import android.graphics.Color;
+import android.media.AudioFormat;
+import android.media.AudioRecord;
+import android.media.MediaRecorder;
+import android.os.AsyncTask;
 import android.os.Bundle;
+import android.os.Environment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentTransaction;
 import android.support.v7.app.AppCompatActivity;
@@ -12,6 +17,14 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.Toast;
 
+import java.io.BufferedOutputStream;
+import java.io.DataOutputStream;
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.OutputStream;
+import java.util.Arrays;
+
 
 public class MainActivity extends AppCompatActivity {
     private static final String TAG = MainActivity.class.getSimpleName();
@@ -20,6 +33,9 @@ public class MainActivity extends AppCompatActivity {
     int rec_btn_count = 0, play_btn_count =0;
     GraphFragment graphFragment = new GraphFragment();
     ListFragment listFragment = new ListFragment();
+    AudioRecordClass audioRecordClass;
+    Boolean isRecording = false;
+    public static int temp;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -50,21 +66,17 @@ public class MainActivity extends AppCompatActivity {
             Rec.setOnClickListener(new View.OnClickListener() {
                 @Override
                 public void onClick(View v){
-
+                    audioRecordClass = new AudioRecordClass();
                     if (rec_btn_count == 0){
                         //RECORD Button
                         Log.d("VIVZ", "Clicked - Record");
                         Rec.setText("STOP");
                         Rec.setTextColor(Color.parseColor("#ff0000"));
                         Play.setEnabled(false);
+                        isRecording = true;
+                        audioRecordClass.execute();
 
-                        Thread recordThread = new Thread(new Runnable(){
-                            @Override
-                            public void run() {
-                                AudioRecordClass.startRecord();
-                            }
-                        }); // End of record Thread
-                        recordThread.start();
+
                         Toast.makeText(getApplicationContext(), "Recording started", Toast.LENGTH_SHORT).show();
                         rec_btn_count =1;
                     }
@@ -72,7 +84,7 @@ public class MainActivity extends AppCompatActivity {
                     else if (rec_btn_count == 1){
                         //STOP Button
                         Log.d("VIVZ", "Clicked - Stop");
-                        AudioRecordClass.stopRecord();
+                        isRecording = false;
                         Rec.setText("RECORD");
                         Rec.setTextColor(Color.parseColor("#000000"));
                         Play.setEnabled(true);
@@ -115,5 +127,149 @@ public class MainActivity extends AppCompatActivity {
         }//End of Play Button
 
     }// End of onCreate()
+    public class AudioRecordClass extends AsyncTask{
+
+
+        public Boolean recording = true;
+
+        @Override
+        protected Object doInBackground(Object[] objects) {
+            startRecord();
+            return null;
+        }
+
+        public void startRecord(){
+            Log.d("VIVZ", "Thread - Start record");
+        /* WHOLE PROCESS EXPLAINED IN BRIEF HERE:
+            1.Create a file to store that data values that comes from the mic.
+            2. Fix the bufferSize and AudioRecord Object.(Will be later in detail later).
+            3.In java the data comes in the form of bytes-bytes-bytes-and so on.
+            4.In the file that we have created we can store the same byte recieved.
+            5.But as we have to use 16 bit PCM ENCODING SYSTEM(Quantitaion), We cannot store the data in Byte form.
+            6.Thus we convert the data in short datatype and then store the array of short into the file.
+            7. short(16 bit) = 2*byte(8-bit)
+            8.And here we have used file to store the audio value from Mic and used the same file to play the Audio.
+            9.We store the data in file as Short-Short-Short(array of short) and fetch the data in same way to fetch.
+            10.But simply saying we do not needed to store and fetch from file for recording and playing for ONCE.
+            11.for that purpose , we can use the array of short datatype
+            12. Another thing is when we try to open the file via a text editor (notepad /notepad++ used by us) we cannot read
+                the actual data(short datatype) that we have store in that file.Because we have stored 16bit-16bit-16bit----
+                and most of the text editor use UTF-8 encoding which is 32-bit.
+            13.Thus to read the data we have to store it using int datatypte . int-int-int
+            14.And in this case we have to name the extension as (.txt).But when we store and fetch the data ourselves to mic and speaker
+                respectively, the extension does not matter at all . To show that I have craeted Three File
+                ONE- as extension Sound.pcm
+                Two- as extension Sound.haha
+                Three- as extension Sound.txt
+             15. AND MOST IMPORTANT THING TO REMEMBER :- OUR AMPLITUDE IS REPRESENTED BY 16 bit. SO WE USE SHORT
+         */
+
+
+            File filePcm = new File(Environment.getExternalStorageDirectory(),"Sound.pcm");
+            File fileHaha = new File(Environment.getExternalStorageDirectory(),"Sound.haha");
+            File fileTxt = new File(Environment.getExternalStorageDirectory(),"Sound.txt");
+       /*  -Above Three are Three different files as discussed above. In first two the files we pass the Array of short as the data
+            to be stored and similarly fetch the data in same way.This is to that the extension does not effect.
+           -And the Third kind of file stores tha data in integer form and has extension .txt so that text Editor(UFT-8) can
+            open and understahnd and show the data.PLEASE, NOTE THAT EXTENSION DOES AFFECT HERE.
+*/
+
+            try {
+                filePcm.createNewFile();
+                fileHaha.createNewFile();
+                fileTxt.createNewFile();
+
+                // Mechanism to store fetch data from mic and store it.
+                OutputStream outputStream = new FileOutputStream(fileHaha);
+                BufferedOutputStream bufferedOutputStream = new BufferedOutputStream(outputStream);
+                DataOutputStream dataOutputStream = new DataOutputStream(bufferedOutputStream);
+
+                // Mechanism to store fetch data from mic and store it.
+                OutputStream outputStream1 = new FileOutputStream(filePcm);
+                BufferedOutputStream bufferedOutputStream1 = new BufferedOutputStream(outputStream1);
+                DataOutputStream dataOutputStream1 = new DataOutputStream(bufferedOutputStream1);
+
+                // Mechanism to store fetch data from mic and store it.
+                OutputStream outputStream2 = new FileOutputStream(fileTxt);
+                BufferedOutputStream bufferedOutputStream2 = new BufferedOutputStream(outputStream2);
+                DataOutputStream dataOutputStream2 = new DataOutputStream(bufferedOutputStream2);
+
+            /*Call the static class of Audio Record to get the Buffer size in Byte that can handle the Audio data values
+                based on our SAMPLING RATE (44100 hz or frame per second in our case)
+             */
+                int minBufferSize = AudioRecord.getMinBufferSize(44100,
+                        AudioFormat.CHANNEL_IN_DEFAULT,
+                        AudioFormat.ENCODING_PCM_16BIT);
+
+                // The array short that will store the Audio data that we get From the mic.
+                short[] audioData = new short[minBufferSize];
+                float[] audioFloats= new float[audioData.length];
+
+                //Create a Object of the AudioRecord class with the required Samplig Frequency(44100 hz)
+                AudioRecord audioRecord = new AudioRecord(MediaRecorder.AudioSource.MIC,
+                        44100,
+                        AudioFormat.CHANNEL_IN_DEFAULT,
+                        AudioFormat.ENCODING_PCM_16BIT,
+                        minBufferSize);
+
+            /* object of the AudioRecord class calls the startRecording() function so that every is ready and the data
+                can be fetch from mic-buffer-our array of short(audioData)
+             */
+                audioRecord.startRecording();
+                //GraphFragment gF = new GraphFragment();
+                // it means while the user have  not pressed the STOP Button
+                while(isRecording){
+
+                /* numberOfShort=minBufferSize/2
+                   Actually what is happening is the minBufferSize(8 bit Buffer) is being converted to numberOfShort(16 bit buffer)
+                   AND THE MOST IMPORTANT PART IS HERE:- The actual value is being store here in the audioData array.
+                 */
+                    int numberOfShort = audioRecord.read(audioData, 0, minBufferSize);
+
+                /*This is part where we store that data to our 3 different files.
+                   For now I have used (.haha) and (.txt)
+                 */
+                    for(int i = 0; i < numberOfShort; i++){
+                        dataOutputStream.writeShort(audioData[i]); // Store in Sound.haha file as short-short-short--
+                        dataOutputStream1.writeShort(audioData[i]);
+
+                        temp = (int)audioData[i];//Convert the short to int to store in txt file
+                        //GraphFragment.graph_height=temp;
+                        audioFloats[i] = ((float)Short.reverseBytes(audioData[i])/0x8000);
+                        dataOutputStream2.writeInt(temp);//Store in Sound.txt as int-int-int--
+                    }
+
+                }
+
+                /** FFT calculation part **/
+
+//            float[] fft_input = new float[8];
+//            for(int i=0;i<8;i++){
+//                fft_input[i] = audioFloats[i];
+//            }
+//            FFT fft_object= new FFT(fft_input);
+                /*double[] fftAbsoluteOutput= FftOutput.callMainFft(audioFloats);
+                System.out.println("absolute value: "+ Arrays.toString(fftAbsoluteOutput));
+                double[] frequency = FrequencyValue.getFrequency(fftAbsoluteOutput);
+                System.out.println("Frequency value: "+ Arrays.toString(frequency));
+                System.out.println(fftAbsoluteOutput.length);
+                System.out.println(frequency.length);*/
+                audioRecord.stop();
+
+                System.out.println("Audio Data: "+ Arrays.toString(audioData));
+                dataOutputStream.close();
+                dataOutputStream1.close();
+                dataOutputStream2.close();
+
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+
+        }
+
+        public void stopRecord(){
+            recording = false;
+        }
+    }
 
 }//End of MainActivity
